+++ tests/test_anime.py
@@ -34,7 +34,7 @@ def test_request_anime_bad(client):
     xml = testlib.load_text('anime_bad.xml')
     with mock.patch('mir.anidb.api.httpapi_request') as request:
         request.return_value = testlib.FakeResponse(xml)
-        with pytest.raises(ValueError):
+        with pytest.raises(anime.MissingElementError):
             anime.request_anime(client, 22)
     request.assert_called_once_with(client, request='anime', aid=22)